pyproject.toml
[project]
name = "explodable"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "anthropic>=0.89.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",
    "psycopg[binary,pool]>=3.2.0",
    "pgvector>=0.3.0",
    "pydantic>=2.0.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "fastapi>=0.115.0",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "structlog>=24.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "datasketch>=1.6.0",
    "textstat>=0.7.0",
    "openai>=1.0.0",
    "exa-py>=1.0.0",
    "langgraph-checkpoint-postgres>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
